Misc test fixes (#9902)
* Fix up ndarray creation with empty databuffers (length == 0)
Fix up cbow inference array creation (wrong vector usage in locked words creation)
Use shape info from passed in copy array constructor

* Add UDFs ADR
Add new op type UDF
Update flatbuffers to reflect new changes
Add annotation scanning for udfs
Add op execution for UDFS

* Add UDFs ADR
Add new op type UDF
Update flatbuffers to reflect new changes
Add annotation scanning for udfs
Add op execution for UDFS

* Add UDFs ADR
Add new op type UDF
Update flatbuffers to reflect new changes
Add annotation scanning for udfs
Add op execution for UDFS
Add UDF and test for training

* Add UDFs ADR
Add new op type UDF
Update flatbuffers to reflect new changes
Add annotation scanning for udfs
Add op execution for UDFS
Add UDF and test for training

* Add UDFs ADR
Add new op type UDF
Update flatbuffers to reflect new changes
Add annotation scanning for udfs
Add op execution for UDFS
Add UDF and test for training
Licenses

* Add UDFs ADR
Add new op type UDF
Update flatbuffers to reflect new changes
Add annotation scanning for udfs
Add op execution for UDFS
Add UDF and test for training
Licenses
Update cuda

* Add UDFs ADR
Add new op type UDF
Update flatbuffers to reflect new changes
Add annotation scanning for udfs
Add op execution for UDFS
Add UDF and test for training
Licenses
Update cuda

* Add UDFs ADR
Add new op type UDF
Update flatbuffers to reflect new changes
Add annotation scanning for udfs
Add op execution for UDFS
Add UDF and test for training
Licenses
Update cuda

* Add UDFs ADR
Add new op type UDF
Update flatbuffers to reflect new changes
Add annotation scanning for udfs
Add op execution for UDFS
Add UDF and test for training
Licenses
Update cuda

* Add UDFs ADR
Add new op type UDF
Update flatbuffers to reflect new changes
Add annotation scanning for udfs
Add op execution for UDFS
Add UDF and test for training
Licenses
Update cuda
Remove extra

* Add UDFs ADR
Add new op type UDF
Update flatbuffers to reflect new changes
Add annotation scanning for udfs
Add op execution for UDFS
Add UDF and test for training
Licenses
Update cuda
Remove extra generated files

* Add extra case for xw_plus_b
Fix scalar bool eps
Misc test fixes

+++ platform-tests/src/test/java/org/datavec/jdbc/impl/JDBCRecordReaderTest.java
@@ -62,7 +62,8 @@ public class JDBCRecordReaderTest {
 
     @BeforeEach
     void setUp() throws Exception {
-        FileUtils.forceDelete(new File(dbName));
+        if(new File(dbName).exists())
+            FileUtils.forceDelete(new File(dbName));
         dataSource = new EmbeddedDataSource();
         dataSource.setDatabaseName(dbName);
         dataSource.setCreateDatabase("create");
@@ -249,7 +250,7 @@ public class JDBCRecordReaderTest {
             // varchar to text
             assertEquals(Text.class, item.get(6).getClass());
             assertEquals(DoubleWritable.class, // float to double (derby's float is an alias of double by default)
-            item.get(7).getClass());
+                    item.get(7).getClass());
             // real to float
             assertEquals(FloatWritable.class, item.get(8).getClass());
             // decimal to double
